refactor to use ListCrudRepository
